Sponsored block modifications on Industry Week
Image size for sponsored/editorial advertisements down to 180, spacing adjusted to compensate for smaller image.

On QMN, images moved to the left for editorial content, and on the right for advertisements.

I think they’re restructuring their newsletters in the next round of changes, so I’m holding off on templatizing these tweaks until we know for sure what they’re sticking with.