Extended “right-rail” ad support to additional eNLs
Per Megan Garcia via CS-3953, these are all the newsletters she knows of that will have the 160x600 ad on the right with text.

Transferred that support into a component for reuse, cleaned up e styling and added valign support to core-table.marko, otherwise the column will center vertically unless there’s enough content on the left to exceed the height of the ad on the right.