Stealth Link added globally
Per Jackie via https://southcomm.atlassian.net/browse/CS-4245

I added the newsletter name as the hidden link text just so the <a> is wrapped around something.  Hidden text still appears in preview below the subject line of some email clients (like outlook) so I wanted it to be somewhat relevent.

Omeda’s documentation states that this can be placed anywhere within the <body> tag, and that the location of the stealth link will not affect the link’s functionality, so I placed it just after the “View Online” text.