Allow brandNavLinks to still be passed on top of config
That way, Home, or any other links that need to be added that aren’t customizable thorugh the interface can be done on short-notice.  Home shouldn’t really need to be customized by the user, so I left that statically-set.

`subscribe` will need to be added to the fragment once it’s been merged in base-cms, I left that out for now so it wouldn’t error.