Added dateColor and descriptionColor for white backgrounds
There were a handful of newsletters that would have a black date and gray description color, with a custom newsletter name color.

This will check to see if the background color hex includes “#fff” and set the date color and description color to something other than the font color.  For newsletters with colored backgrounds, the date and descriptions are the same color as the newsletter name.

This won’t work if the user sets the background color to “white” or if they do something like ‘#fff322’ but all variations of ‘#fffXXX” are bright yellow, so these color overrides will still look nice.